+++ pyvela/pyvela/pyvela_plot.py
@@ -15,6 +15,7 @@ import numpy as np
 from astropy import units as u
 from pint import DMconst, dmu
 from pint.models import get_model
+from scipy.stats import kstest
 
 
 def get_param_plot_mask(
@@ -79,6 +80,15 @@ def get_psrname(result_dir: str) -> Optional[str]:
         return model["PSR"].value
     except:
         return ""
+
+
+def get_pepoch(result_dir: str) -> float:
+    with open(f"{result_dir}/summary.json") as summary_file:
+        summary = json.load(summary_file)
+
+    parfile = summary["input"]["par_file"]
+    model = get_model(f"{result_dir}/{parfile}", allow_tcb=True, allow_T2=True)
+    return model["PEPOCH"].value
 
 
 def plot(
@@ -207,6 +217,42 @@ def plot(
         ax1.axhline(0, ls="dotted", color="k")
 
     ax.set_xlabel("MJD - PEPOCH")
+
+    ax3 = plt.subplot(5, 3, 2)
+    ax3.set_ylim((0, 1))
+    ax3.axis("off")
+    pepoch = get_pepoch(result_dir)
+    if not wb:
+        weights = 1 / terr**2
+        wrms = np.sqrt(np.average(tres_w**2, weights=weights))
+        ks = kstest(tres_w / terr, "norm", args=(0, 1))
+        ax3.text(
+            0,
+            0,
+            f"""
+            No of TOAs = {len(tres_w)}
+            MJD Range = {int(min(mjds) + pepoch)} — {int(max(mjds) + pepoch)}
+            Wrms time resids = {wrms:.2e} s
+            KS test p-value = {ks.pvalue:.2e}
+            """,
+        )
+    else:
+        weights_t = 1 / terr**2
+        weights_d = 1 / derr**2
+        wrms_t = np.sqrt(np.average(tres_w**2, weights=weights_t))
+        wrms_d = np.sqrt(np.average(dres_w**2, weights=weights_d))
+        ks = kstest(np.append(tres_w / terr, dres_w / derr), "norm", args=(0, 1))
+        ax3.text(
+            0,
+            0,
+            f"""
+            No of TOAs = {len(tres_w)}
+            MJD Range = {int(min(mjds))} -- {int(max(mjds))}
+            Wrms time resids = {wrms_t:.2e} s
+            Wrms DM resids = {wrms_d:.2e} pc/cm^3
+            KS test p-value = {ks.pvalue:.2e}
+            """,
+        )
 
     psrname = get_psrname(result_dir)
     if psrname is not None: