+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Display default arguments in help messages of all scripts
 - Work stealing in `pyvela` CI tests
 - Optionally specify Gaussian priors for analytically marginalized timing model parameters
+- Weighted residual RMS and KS test p-value in `pyvela-plot`
 ## Changed
 - Split the `SPNTA.save_results()` method into two -- the pre-analysis details are now saved using `SPNTA.save_pre_analysis_summary()` before running MCMC.
 ## Fixed